Extracting runFile function into a separate file. #370
