+++ bin/taiko.js
@@ -1,12 +1,11 @@
 #! /usr/bin/env node
 
-const path = require('path');
-const util = require('util');
+const { runFile } = require('./runFile');
 const fs = require('fs');
 const program = require('commander');
 const taiko = require('../lib/taiko');
 const repl = require('../lib/repl');
-const { removeQuotes, symbols, isTaikoRunner } = require('../lib/util');
+const { isTaikoRunner } = require('../lib/util');
 const devices = require('../lib/device').default;
 let repl_mode = false;
 
@@ -25,37 +24,6 @@ async function exitOnUnhandledFailures(e) {
 
 process.on('unhandledRejection', exitOnUnhandledFailures);
 process.on('uncaughtException', exitOnUnhandledFailures);
-
-function runFile(file, observe, observeTime) {
-    const realFuncs = {};
-    for (let func in taiko) {
-        realFuncs[func] = taiko[func];
-        if (realFuncs[func].constructor.name === 'AsyncFunction') global[func] = async function () {
-            let res, args = arguments;
-            if (func === 'openBrowser') {
-                if (args['0']) { args['0'].headless = !observe; args[0].observe = observe; args['0'].observeTime = observeTime; }
-                else args = [{ headless: !observe, observe: observe, observeTime: observeTime }];
-            }
-            res = await realFuncs[func].apply(this, args);
-            if (res.description) {
-                res.description = symbols.pass + res.description;
-                console.log(removeQuotes(util.inspect(res.description, { colors: true }), res.description));
-            }
-            return res;
-        };
-        else global[func] = function () {
-            return realFuncs[func].apply(this, arguments);
-        };
-        require.cache[path.join(__dirname, 'taiko.js')].exports[func] = global[func];
-    }
-    const oldNodeModulesPaths = module.constructor._nodeModulePaths;
-    module.constructor._nodeModulePaths = function () {
-        const ret = oldNodeModulesPaths.apply(this, arguments);
-        ret.push(__dirname);
-        return ret;
-    };
-    require(path.resolve(file).slice(0, -3));
-}
 
 function validate(file) {
     if (!file.endsWith('.js')) {
@@ -84,7 +52,7 @@ program
        taiko <file> [options]`)
     .option(
         '-o, --observe',
-        ` enables headful mode and runs script with 3000ms delay by default.
+        `enables headful mode and runs script with 3000ms delay by default.
         \t\t\tpass --wait-time option to override the default 3000ms\n`
     )
     .option(
